backend/requirements.txt
flask
flask-cors
pymongo
bcrypt
transformers
torch
pillow
